// Web/src/app/models/CropVM.ts
export class CropVM{
    ID_crop: number;
    name: string;
    variety_name: string;
    description: string;
    type: string;
    avg_yield_acre: number;
    avg_size?: number;
    avg_color: string;
    maturity_cycle: string;
    crop_cycle: string;
    crop_season: string;
    kc_init?: number;
    kc_mid?: number;
    kc_end?: number;
    stage_1: string;
    stage_2: string;
    stage_3: string;
    stage_4: string;
    stage_5: string;
    stage_6: string;
    stage_7: string;
    stage_8: string;
    stage_9: string;
    stage_10: string;    
    company: number;
}

export class SearchCropVM{
    TotalRows: number;
    ID_crop: number;
    name: string;
    variety_name: string;
    description: string;
    type: string;
    avg_yield_acre: number;
    avg_size?: number;
    avg_color: string;
    maturity_cycle: string;
    crop_cycle: string;
    crop_season: string;
    kc_init?: number;
    kc_mid?: number;
    kc_end?: number;
    stage_1: string;
    stage_2: string;
    stage_3: string;
    stage_4: string;
    stage_5: string;
    stage_6: string;
    stage_7: string;
    stage_8: string;
    stage_9: string;
    stage_10: string;
    company: number;
}